Made gallery page dynamic without ssr for redux support
